+++ CHANGELOG.md
@@ -6,6 +6,8 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 
 ## [x.x.x]
 ### Fixed
+- Include markdown for release autodeploy docs
+- Use standard inheritance model in CLinVar (https://ftp.ncbi.nlm.nih.gov/pub/GTR/standard_terms/Mode_of_inheritance.txt)
 - Make MitoMap link work for hg38 again
 ### Added
 - Add link to HmtVar for mitochondrial variants (if VCF is annotated with HmtNote)